+++ Assets/Materials/Shader Graphs_Paintable.mat.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 12ccca4edd3ea5c419e87d2b2f889132
+guid: c52a95f0ba37e8a4b871b08519e1a057
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 8400000
+  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
