+++ Assets/Prefabs/PlayerCapsule.prefab
@@ -188,6 +188,7 @@ GameObject:
   - component: {fileID: 8616685848737228376}
   - component: {fileID: 2308672497553814287}
   - component: {fileID: 4629075709156929370}
+  - component: {fileID: 7464195706511024604}
   m_Layer: 8
   m_Name: PlayerCapsule
   m_TagString: Player
@@ -433,3 +434,18 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   _holdPosition: {fileID: 4938015035779606454}
+--- !u!114 &7464195706511024604
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8616685848737228372}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 54cbcbdee6551e64c935cf348d1566e7, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _radius: 0.1
+  _strength: 1
+  _hardness: 1
