+++ Assets/Prefabs/paintable box.prefab
@@ -1,0 +1,108 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1001 &1814324230733249588
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: -8679921383154817045, guid: c20df6d6a40b614458e0ee74c95aeb3e, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 2.2662945
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: c20df6d6a40b614458e0ee74c95aeb3e, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: c20df6d6a40b614458e0ee74c95aeb3e, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 4.044089
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: c20df6d6a40b614458e0ee74c95aeb3e, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 0.7071067
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: c20df6d6a40b614458e0ee74c95aeb3e, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: -0.7071068
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: c20df6d6a40b614458e0ee74c95aeb3e, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: c20df6d6a40b614458e0ee74c95aeb3e, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: c20df6d6a40b614458e0ee74c95aeb3e, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: c20df6d6a40b614458e0ee74c95aeb3e, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: c20df6d6a40b614458e0ee74c95aeb3e, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7511558181221131132, guid: c20df6d6a40b614458e0ee74c95aeb3e, type: 3}
+      propertyPath: 'm_Materials.Array.data[0]'
+      value: 
+      objectReference: {fileID: 2100000, guid: c52a95f0ba37e8a4b871b08519e1a057, type: 2}
+    - target: {fileID: 919132149155446097, guid: c20df6d6a40b614458e0ee74c95aeb3e, type: 3}
+      propertyPath: m_Name
+      value: paintable box
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: c20df6d6a40b614458e0ee74c95aeb3e, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 966702705701234759}
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: c20df6d6a40b614458e0ee74c95aeb3e, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 8752674415177523664}
+  m_SourcePrefab: {fileID: 100100000, guid: c20df6d6a40b614458e0ee74c95aeb3e, type: 3}
+--- !u!1 &1579830634861298021 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 919132149155446097, guid: c20df6d6a40b614458e0ee74c95aeb3e, type: 3}
+  m_PrefabInstance: {fileID: 1814324230733249588}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &966702705701234759
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1579830634861298021}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c47ddcb67536d7440bb603f3b98809af, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _paintableShader: {fileID: 4800000, guid: b8d76d4e901440b2974df15a4e39dcd5, type: 3}
+--- !u!65 &8752674415177523664
+BoxCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1579830634861298021}
+  m_Material: {fileID: 0}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_LayerOverridePriority: 0
+  m_IsTrigger: 0
+  m_ProvidesContacts: 0
+  m_Enabled: 1
+  serializedVersion: 3
+  m_Size: {x: 2, y: 1.9999998, z: 1.9999998}
+  m_Center: {x: 0, y: 0, z: 0}
