+++ Assets/Prefabs/paintable box.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5a539b98f9d217b4d96fd37193885820
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
